requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dependency-injector==4.46.0
event_core @ git+https://github.com/axwhyzee/multi-modal-retrieval-event-core.git
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.37
langchain-text-splitters==0.3.6
langsmith==0.3.9
numpy==2.2.3
opencv-python==4.11.0.86
orjson==3.10.15
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.4
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-toolbelt==1.0.0
scenedetect==0.6.5.2
sniffio==1.3.1
tenacity==9.0.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
zstandard==0.23.0
